+++ androiddemo/src/main/java/com/ty/animation/DecelerateInterpolator.java
@@ -40,7 +40,7 @@ public class DecelerateInterpolator implements Interpolator {
     public float getInterpolation(float input) {
         float result;
         if (mFactor == 1.0f) {
-            result = (float)(1.0f - (1.0f - input) * (1.0f - input));
+            result = 1.0f - (1.0f - input) * (1.0f - input);
         } else {
             result = (float)(1.0f - Math.pow((1.0f - input), 2 * mFactor));
         }
